feat(loading handler): generic hook
