+++ packages/mantine/src/hooks/index.ts
@@ -1,2 +1,3 @@
-export * from './useParentHeight';
 export * from './useControlledList';
+export * from './useLoadingHandler';
+export * from './useParentHeight';
